Add chart with rankings to View Issue page form
Use the chart.js library from MantisGraph, to avoid having to bundle it
again.

Note that this will only work with MantisBT < 2.8.0, after that
MantisGraph::resources() method restricted loading of chart.js to the
plugin's own pages.

The MantisGraph plugin will need an update to expose a function to load
the chart.js resources to enable this functionality with recent versions
of MantisBT.